@@ -6,13 +6,13 @@
 
     <groupId>ru.tioplaya</groupId>
     <artifactId>PreventFallLite</artifactId>
-    <version>1.1</version>
+    <version>1.2</version>
     <packaging>jar</packaging>
 
     <name>PreventFallLite</name>
 
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>11</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
 
